Take into account the config file last write time
The timestamping mechanism was previously only caching/looking at the last write time of source image assets to determine if resizing should happen again.  However, there is another scenerio (and probably more likely) in which the source image assets don’t change, but the config file does which should cause the images to be regenerated anyway.  This should take the last write time of the config file into account too and regenerate output if the config file was changed.